refactor: group math helpers
