+++ src/app/common/math/calculate-averages.spec.ts
