+++ src/app/common/math/calculate-averages.ts
