+++ src/app/common/utils.spec.ts
@@ -1,9 +1,7 @@
-import BigNumber from 'bignumber.js';
-
-import { getTicker, sumNumbers } from '@app/common/utils';
+import { getTicker } from '@app/common/utils';
 import { extractPhraseFromString } from '@app/common/utils';
 
-import { countDecimals } from './utils';
+import { countDecimals } from './math/helpers';
 
 describe(countDecimals.name, () => {
   test('that it returns 0 when given an integer', () => expect(countDecimals(100)).toEqual(0));
@@ -105,28 +103,5 @@ describe(extractPhraseFromString.name, () => {
   test('Extract secret key from seed copied from v3 wallet', () => {
     const key = extractPhraseFromString(SECRET_KEY_COPIED_FROM_v3);
     expect(key).toEqual(SECRET_KEY_COPIED_FROM_V3_FORMATTED);
-  });
-});
-
-const cases = [
-  {
-    sums: [1, 2, 3],
-    expectedResult: new BigNumber(6),
-  },
-  {
-    sums: [0.1, 0.2],
-    expectedResult: new BigNumber(0.3),
-  },
-  {
-    sums: [Number.MAX_SAFE_INTEGER, 1],
-    expectedResult: new BigNumber('9007199254740992'),
-  },
-];
-
-describe(sumNumbers.name, () => {
-  describe.each(cases)('Sum example', ({ sums, expectedResult }) => {
-    test('sum of ' + sums.toString(), () => {
-      expect(sumNumbers(sums).toString()).toEqual(expectedResult.toString());
-    });
   });
 });
