+++ src/app/pages/send/ordinal-inscription/select-inscription-coins.spec.ts
@@ -1,4 +1,4 @@
-import { sumNumbers } from '@app/common/utils';
+import { sumNumbers } from '@app/common/math/helpers';
 
 import { selectInscriptionTransferCoins } from './select-inscription-coins';
